# sql: don't decode the first key TableID/IndexID