The RowFetcher is never called on a span that includes more than one
top-level index, so it's safe to staticly determine the length of the
keys' table id / index id prefix and skip it during index key decode.

```
name                                old time/op    new time/op    delta
WideTableIgnoreColumns/Cockroach-8    15.2ms ± 1%    15.0ms ± 1%  -1.28%  (p=0.002 n=10+10)
```

Partial solution to #21425.

Release note (performance improvement): do less work during key decode