@RaduBerinde 2018-01-17
Review status: 0 of 2 files reviewed at latest revision, 2 unresolved discussions, some commit checks failed.

---

*[pkg/sql/sqlbase/table.go, line 773 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/21498#-L31DHrg1VF_XwefQ0mt:-L31DHrg1VF_XwefQ0mu:b-ifxqi2) ([raw file](https://github.com/cockroachdb/cockroach/blob/4ef95d75c0dfdb9090674fdf9b91431c9388c450/pkg/sql/sqlbase/table.go#L773)):*
> ```Go
> 	key []byte,
> ) (remainingKey []byte, matches bool, _ error) {
> 	knownPrefixLength := len(MakeIndexKeyPrefix(desc, index.ID))
> ```

Is this version of the function slower now (because we allocate a key)?

---

*[pkg/sql/sqlbase/table.go, line 777 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/21498#-L31D538Osn_DP0Q_G3k:-L31D539SUEUYdp6VwAg:b-qd0lgd) ([raw file](https://github.com/cockroachdb/cockroach/blob/4ef95d75c0dfdb9090674fdf9b91431c9388c450/pkg/sql/sqlbase/table.go#L777)):*
> ```Go
> }
> 
> func DecodeIndexKeyWithPrefixLen(
> ```

Why not pass `key[knownPrefixLength:]` to this function instead of passing the full key and a length we always skip

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/21498)*
<!-- Sent from Reviewable.io -->