@jordanlewis 2018-01-17
Review status: 0 of 2 files reviewed at latest revision, 2 unresolved discussions, some commit checks failed.

---

*[pkg/sql/sqlbase/table.go, line 773 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/21498#-L31DHrg1VF_XwefQ0mt:-L31ED5hKrt-R6-zQXhy:b-ww154z) ([raw file](https://github.com/cockroachdb/cockroach/blob/4ef95d75c0dfdb9090674fdf9b91431c9388c450/pkg/sql/sqlbase/table.go#L773)):*
<details><summary><i>Previously, RaduBerinde wrote…</i></summary><blockquote>

Is this version of the function slower now (because we allocate a key)?
</blockquote></details>

Yes, but it's only used by tests and another non-hotpath function.

---

*[pkg/sql/sqlbase/table.go, line 777 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/21498#-L31D538Osn_DP0Q_G3k:-L31EIOzEo41_wb5-i1h:baxmz27) ([raw file](https://github.com/cockroachdb/cockroach/blob/4ef95d75c0dfdb9090674fdf9b91431c9388c450/pkg/sql/sqlbase/table.go#L777)):*
<details><summary><i>Previously, RaduBerinde wrote…</i></summary><blockquote>

Why not pass `key[knownPrefixLength:]` to this function instead of passing the full key and a length we always skip
</blockquote></details>

Yeah that's a better idea. I'll redo this a little bit.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/21498)*
<!-- Sent from Reviewable.io -->