@RaduBerinde 2018-01-17
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

---

Review status: 0 of 3 files reviewed at latest revision, all discussions resolved, all commit checks successful.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/21498#-:-L34t4CYMfb1jS2ACK1S:bnfp4nl)*
<!-- Sent from Reviewable.io -->